Rewrite Json output for fixup to make it more readable
