+++ src/main/scala/com/gu/tagdiffer/fixup/Json.scala
@@ -1,0 +1,5 @@
+package com.gu.tagdiffer.fixup
+
+trait Json {
+  implicit val
+}
